+++ CSETWebNg/src/app/assessment/questions/category-block/category-block.component.html
@@ -1,6 +1,6 @@
 <!-- Question Group Heading / Category-->
-<div *ngIf="c.ShowOverrideHeader" class="group-heading" [class.display-none]="!c.Visible">
-  {{ c.Symbol_Name + ' - ' + c.ComponentName }} ABC
+<div *ngIf="c.IsOverride" class="group-heading" [class.display-none]="!c.Visible">
+  {{ c.Symbol_Name + ' - ' + c.ComponentName }}
 </div>
 
 <!-- Question Group Heading / Category-->
